add paired stats to DataCollection (and move data collections to their own file)
