+++ setup.cfg
@@ -4,7 +4,9 @@ description-file = readme.md
 [pytest]
 pep8ignore =
     E501
-    wqio/tests/test_features.py E241
-    wqio/tests/test_ros.py E241
     wqio/features.py E131
+    wqio/datacollections.py E131
+    wqio/tests/test_features.py E241
+    wqio/tests/test_datacollections.py E241
+    wqio/tests/test_ros.py E241
     docs/* ALL
